Highlight: move utils to their own module
